move implemenation of snippet search to a redux action
create SEARCH_SNIPPET action and update corresponding reducer
use fuse.js search within SEARCH_SNIPPET action creator and initialize search object with each snippet fetch
use snippet data from props instead of copying to state in SnippetList component
