+++ client/src/components/snippets/SnippetList.tsx
@@ -3,6 +3,7 @@ import { connect } from 'react-redux';
 import { Link } from 'react-router-dom';
 import {
   fetchSnippets, 
+  searchSnippetList, 
   fetchUsers,
 } from '../../store/actions';
 import {
@@ -13,7 +14,6 @@ import {
 import {
   isEmpty,
 } from '../../lib/lib';
-import Fuse from 'fuse.js';
 import FilterSnippetForm from './filterSnippetForm';
 
 
@@ -21,6 +21,7 @@ import FilterSnippetForm from './filterSnippetForm';
 interface SnippetListProps {
   fetchSnippets: (filters?: any) => void;
   fetchUsers: (selectedTeam: any) => void;
+  searchSnippetList: (searchText: any) => void;
 
   snippets: Array<Snippet> | null;
   user: User | null;
@@ -29,35 +30,16 @@ interface SnippetListProps {
 }
 
 interface SnippetListState {
-  snippets: Array<Snippet>;
   searchText: string;
 }
 
-interface FuseOptions {
-  shouldSort: boolean;
-  tokenize: boolean;
-  keys: Array<string>;
-}
-
-const fuseOpts: FuseOptions = {
-  shouldSort: true,
-  tokenize: true,
-  keys: [
-    'title',
-    'content',
-    'description',
-    'ownerName', 
-  ],
-};
 
 class SnippetList extends Component<SnippetListProps, SnippetListState> {
-  fuse: Fuse<Snippet, FuseOptions> | null = null;
   
   constructor(props: SnippetListProps) {
     super(props);
 
     this.state = {
-      snippets: props.snippets ? props.snippets : [],
       searchText: '',
     };
   }
@@ -73,13 +55,6 @@ class SnippetList extends Component<SnippetListProps, SnippetListState> {
         weekSelected: week
       });
       await this.props.fetchUsers(team);
-        
-      if (this.props.snippets) {
-        this.setState({
-          snippets: this.props.snippets,
-        });
-        this.fuse = new Fuse(this.props.snippets, fuseOpts);
-      }
     }
   }
 
@@ -88,16 +63,8 @@ class SnippetList extends Component<SnippetListProps, SnippetListState> {
       <div className='row'>
         <form onSubmit={(e) => {
           e.preventDefault();
-          if (this.fuse) {
-            let results: Array<Snippet> = this.props.snippets || [];
-            if (this.state.searchText.length > 0) {
-              results = this.fuse.search(this.state.searchText);
-            }
-            this.setState({
-              snippets: results,
-            });
-          } else {
-            console.error('Fuse search is not initialized');
+          if (this.state.searchText.length > 0) {
+            this.props.searchSnippetList(this.state.searchText);
           }
         }}>
           <input 
@@ -119,8 +86,8 @@ class SnippetList extends Component<SnippetListProps, SnippetListState> {
   }
 
   renderSnippets() {
-    if (this.state.snippets && this.state.snippets.length > 0) {
-      return this.state.snippets.map( (snippet: any) => {
+    if (this.props.snippets && this.props.snippets.length > 0) {
+      return this.props.snippets.map( (snippet: any) => {
         return (
           <li key={snippet.title} className="collection-item">
             <div>
@@ -191,6 +158,10 @@ function mapStateToProps({
   return { snippets, user, selectedTeam, selectedWeek };
 }
 
-export default connect(mapStateToProps, { fetchSnippets, fetchUsers })(
+export default connect(mapStateToProps, { 
+  fetchSnippets, 
+  searchSnippetList, 
+  fetchUsers 
+})(
   SnippetList
 );
