+++ client/src/store/reducers/snippetReducer.ts
@@ -1,8 +1,10 @@
-import { FETCH_SNIPPETS } from '../types'
+import { FETCH_SNIPPETS, SEARCH_SNIPPETS } from '../types'
 
 const snippetReducer = (state = [], action: any) => {
     switch (action.type) {
         case FETCH_SNIPPETS:
+            return action.payload;
+        case SEARCH_SNIPPETS:
             return action.payload;
         default:
             return state;
