+++ client/src/store/types.ts
@@ -82,6 +82,7 @@ export interface State {
 export const FETCH_USER = 'fetch_user';
 export const FETCH_USERS = 'fetch_users';
 export const FETCH_SNIPPETS = 'fetch_snippets';
+export const SEARCH_SNIPPETS = 'search_snippets';
 export const AUTHORIZE_USER = 'authorize_user';
 export const FETCH_TEAMS = 'fetch_teams';
 export const SELECT_TEAM = 'select_team';
